+++ README.md
@@ -1,3 +1,3 @@
-# java-learning
+# java-com.zetian.study.learning
 Java学习
 ## 简单介绍
