+++ src/com/zetian/study/equals/Manager.java
@@ -1,0 +1,44 @@
+package com.zetian.study.equals;
+
+import com.zetian.study.base.EqualsEmployee;
+
+/**
+ * Description 经理
+ *
+ * @author Zetian Wang
+ * @date 2019/09/26
+ **/
+public class Manager extends EqualsEmployee {
+    private double bonus;
+    public Manager(String name,double salary,int year,int month,int day){
+        super(name,salary,year,month,day);
+        bonus=0;
+    }
+
+    @Override
+    public double getSalary(){
+        double baseSalary = super.getSalary();
+        return baseSalary+bonus;
+     }
+
+     public void setBonus(double bonus){
+        this.bonus=bonus;
+     }
+
+    @Override
+     public boolean equals(Object otherObject){
+        if(!super.equals(otherObject)){ return false;}
+        Manager other = (Manager)otherObject;
+        return bonus == other.bonus;
+     }
+
+    @Override
+     public int hashCode(){
+        return super.hashCode()+17*new Double(bonus).hashCode();
+     }
+
+    @Override
+     public String toString() {
+        return super.toString()+"[bonus="+bonus+"]";
+     }
+}
