--- src/equals/Employee.java
@@ -1,65 +1,0 @@
-package equals;
-
-import java.time.LocalDate;
-import java.util.Objects;
-
-/**
- * Description
- *
- * @author Zetian Wang
- * @date 2019/09/25
- **/
-public class Employee {
-    private String name;
-    private double salary;
-    private LocalDate hireDay;
-
-    public Employee(String name, double salary, int year, int month, int day) {
-        this.name = name;
-        this.salary = salary;
-        hireDay = LocalDate.of(year, month, day);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public double getSalary() {
-        return salary;
-    }
-
-    public LocalDate getHireDay() {
-        return hireDay;
-    }
-
-    public void raiseSalary(double buPercent) {
-        double raise = salary * buPercent / 100;
-        salary += raise;
-    }
-
-    @Override
-    public boolean equals(Object otherObject) {
-        if (this == otherObject) {
-            return true;
-        }
-        if (otherObject == null) {
-            return false;
-        }
-        if (getClass() != otherObject.getClass()) {
-            return false;
-        }
-        Employee other = (Employee) otherObject;
-        return Objects.equals(name, other.name) && salary == other.salary && Objects.equals(hireDay, other.hireDay);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, salary, hireDay);
-    }
-
-    @Override
-    public String toString() {
-        return getClass().getName() + "[name" + name + ",salary" + salary + ",hireDay=" + hireDay + "]";
-    }
-}
-
